Modified LIDAR sensor to include red rays and added some basic code to SenseQt trying to integrate PyQt and PyBullet.
In Sensors.py, retrieve_data() was modified from
the original code to be more modulated. The simulate()
function was also added, which uses PyBullet's
ray_test_batch to get the data whether the ray was hit.
This information is polled every time step in the simulation.
Requirements was modified to remove 2 pyyamls.
In SenseQt, code was added to attempt to integrate
PyBullet's GUI and PyQt's MainWindow. This is resulting
in various errors, and can be improved.