Added documentation for SimulationObjects, Sensors and added a CubeCreator class. Also made QtMap work.
QtMap now works correctly, with the red dot representing
the robot's current position and black dots representing
occupied objects. However, moving objects will cause
the robot map to think that a wall is created. Altered
the grid_size to 50 so it renders much faster and
changed the scaling ratio to 1.