Added very simple robot motion to the racecar and a reset map function to the OGM, which is used whenever the simulation is reset.
The robot motion is done by putting all four wheels
and the steering (front two wheels) into two lists.
Whenever W A S D is pressed, either the force is
applied to each wheel or a steering angle is applied
to the front two wheels. This allows the robot to
navigate the plane.