Created QWidget for dialog and implemented into QtBulletSim. Implemented multiple mappings in sim.
RobotDialog takes in two parameters: parent and csv.
The parent is the QMainWindow, and csv is the path
to a csv file. It then parses the csv file into a list
and iterates through this list whenever the "next"
button is pressed. RobotDialog also has some available
styling, where you can set the background color,
text color, button color, and button hover color.
In QtBulletSim, an instance variable "map_version"
can be incremented to go through each of the different
map types. A button will be made to increment this number.